(CDAP-7428) Not to call getClass() in default constructor.
